Get the form data into an array
The HTML file added the necessary form input fields for now.
In the JavaScript file, an event was added to the button that captures the data from the form and creates a new book from it in the array.
In the displayBooks function, the container's child elements of the DOM node had to be removed so that it does not display books multiple times.
In addition, when adding a book, it is checked that it is unique.